In progress snr conversion to usable double
This update has the statistics for figure 5 for both nb2 and nb3 in nb2_apedits.mlx and nb3_apedits.mlx. The cells containing the snr for each waveform are in snr_waveform_b2 are encolsed by {[snr value]} and I cannot use these values in this form for boxplots or statistics. I am working on resolving this problem.